Restore half-orm-packager as standalone package
- Migrate hop tooling from half-orm core to separate package
- Maintain version alignment with half-orm 0.14.0
- All tests passing (dummy_test.sh validated)

+++ half_orm_packager/patch.py
@@ -5,11 +5,17 @@ import os
 import subprocess
 import sys
 
-import pydash
 import psycopg2
 
-from half_orm_packager import utils
-from half_orm_packager import modules
+from half_orm import utils
+from half_orm.packager import modules
+from half_orm.packager.changelog import Changelog
+
+try:
+    PYTEST_OK = True
+    import pytest
+except ImportError:
+    PYTEST_OK = False
 
 class Patch:
     "The Patch class..."
@@ -18,12 +24,10 @@ class Patch:
     def __init__(self, repo):
         self.__repo = repo
         self.__patches_base_dir = os.path.join(repo.base_dir, 'Patches')
-        self.__changelog_file = os.path.join(self.__repo.base_dir, '.hop', 'CHANGELOG')
-        if not os.path.exists(self.__patches_base_dir):
-            os.makedirs(self.__patches_base_dir)
-        if not os.path.exists(self.__changelog_file):
-            utils.write(self.__changelog_file, f'{self.__repo.database.last_release_s}\n')
-        self.__sequence = self.__get_sequence()
+        if self.__repo.devel:
+            self.__changelog = Changelog(repo)
+            if not os.path.exists(self.__patches_base_dir):
+                os.makedirs(self.__patches_base_dir)
 
     @classmethod
     @property
@@ -31,73 +35,78 @@ class Patch:
         "Returns the levels"
         return cls.__levels
 
-    def __get_sequence(self):
-        """Get the sequence of patches in .hop/CHANGLOG"""
-        return [elt.split()[0] for elt in utils.readlines(self.__changelog_file)]
-
-    def __append_to_changelog(self, release_level):
-        """Update with the release the .hop/CHANGELOG file"""
-        utils.write(self.__changelog_file, f'{release_level}\n', mode='a+')
-
-    def __update_changelog(self, release_level):
-        "Add the commit sha1 to the release in the .hop/CHANGELOG file"
-        out = []
-        for line in utils.readlines(self.__changelog_file):
-            line = line.strip()
-            if line and line.split()[0] != release_level:
-                out.append(line)
-            else:
-                out.append(f'{release_level}\t{self.__repo.hgit.last_commit()}\n')
-        utils.write(self.__changelog_file, '\n'.join(out))
-
-    @property
-    def previous(self):
+    def previous(self, release, index=0):
         "Return .hop/CHANGELOG second to last line."
-        return self.__sequence[-2]
+        return self.__changelog.previous(release, index)
 
     @property
-    def last(self):
-        "Return .hop/CHANGELOG last line"
-        return self.__sequence[-1]
-
-    def __repr(self, release):
-        # pylint: disable=consider-using-f-string
-        return '{major}.{minor}.{patch}'.format(**release)
-
-    @property
-    def __next_possible_releases(self):
+    def __next_releases(self):
+        db_last_release = self.__repo.database.last_release_s
+        ch_last_release = self.__changelog.last_release
+        if db_last_release != ch_last_release:
+            utils.error(
+                f'Last release mismatch between database {db_last_release}'
+                f' and CHANGELOG {ch_last_release}!\n', 1)
         current = dict(self.__repo.database.last_release)
-        next_releases = {}
+        releases_in_dev = self.__changelog.releases_in_dev
+        n_rels = {}
         for level in self.__levels:
-            next_releases[level] = dict(current)
-            next_releases[level][level] = current[level] + 1
-        return next_releases
+            n_rel = dict(current)
+            n_rel[level] = current[level] + 1
+            if level == 'major':
+                n_rel['minor'] = n_rel['patch'] = 0
+            if level == 'minor':
+                n_rel['patch'] = 0
+            next_release_s = f"{n_rel['major']}.{n_rel['minor']}.{n_rel['patch']}"
+            n_rel['in_dev'] = ''
+            if next_release_s in releases_in_dev:
+                n_rel['in_dev'] = '(IN DEV)'
+            n_rels[level] = n_rel
+        return n_rels
 
-    def prep_next_release(self, release_level, message=None):
+    def __assert_main_branch(self):
+        if str(self.__repo.hgit.branch) != 'hop_main':
+            utils.error(
+                'ERROR! Wrong branch. Please, switch to the hop_main branch before.\n', exit_code=1)
+
+    def prep_release(self, release_level, message=None):
         """Returns the next (major, minor, patch) tuple according to the release_level
+
+        The repo must be clean.
 
         Args:
             release_level (str): one of ['patch', 'minor', 'major']
         """
+        if not self.__repo.hgit.repos_is_clean():
+            utils.error('There are uncommited changes. Please commit your changes before using `hop prepare`\n', exit_code=1)
+        if self.__repo.database.last_release_s != self.__changelog.last_release:
+            try:
+                self.__restore_db(self.__changelog.last_release)
+            except FileNotFoundError as exc:
+                utils.error(f'No backup file for release {self.__changelog.last_release}\n{exc}\n', exit_code=1)
+        self.__repo.hgit.checkout_to_hop_main()
+        next_releases = self.__next_releases
         if release_level is None:
             next_levels = '\n'.join(
-                [f"- {level}: {self.__repr(self.__next_possible_releases[level])}"
+                [f"""- {level}: {'{major}.{minor}.{patch} {in_dev}'.format(**next_releases[level])}"""
                 for level in self.__levels])
-            print(f'Next possible releases:\n{next_levels}')
-            release_level = input(f"Release level {self.__levels}? ")
-            if not release_level in self.__levels:
+            print(f'Next releases:\n{next_levels}')
+            next_possible_releases = [elt for elt in self.__levels if not next_releases[elt]['in_dev']]
+            release_level = input(f"Release level {next_possible_releases}? ")
+            if release_level not in next_possible_releases:
                 utils.error(f"Wrong release level ({release_level}).\n", exit_code=1)
-        if str(self.__repo.hgit.branch) != 'hop_main':
-            utils.error(
-                'ERROR! Wrong branch. Please, switch to the hop_main branch before.\n', exit_code=1)
+        elif next_releases[release_level]['in_dev']:
+            utils.error(f'{release_level} is alredy in development!\n', 1)
         next_release = dict(self.__repo.database.last_release)
         next_release[release_level] = next_release[release_level] + 1
         if release_level == 'major':
             next_release['minor'] = next_release['patch'] = 0
         if release_level == 'minor':
             next_release['patch'] = 0
-        # pylint: disable=consider-using-f-string
         new_release_s = '{major}.{minor}.{patch}'.format(**next_release)
+        rel_branch = f'hop_{new_release_s}'
+        if self.__repo.hgit.branch_exists(rel_branch):
+            utils.error(f'{rel_branch} already exists!\n', 1)
         print(f'PREPARING: {new_release_s}')
         patch_path = os.path.join(
             'Patches',
@@ -115,11 +124,11 @@ class Patch:
                     'hop_version': utils.hop_version(),
                     'changelog_msg': changelog_msg,
                 }))
+        self.__changelog.new_release(new_release_s)
         self.__repo.hgit.set_branch(new_release_s)
         print('You can now add your patch scripts (*.py, *.sql)'
             f'in {patch_path}. See Patches/README.')
         modules.generate(self.__repo)
-        self.__append_to_changelog(new_release_s)
 
     def __check_apply_or_re_apply(self):
         """Return True if it's the first time.
@@ -129,34 +138,51 @@ class Patch:
             return 're-apply'
         return 'apply'
 
-    def __backup_file(self, release):
-        backup_dir = os.path.join(self.__repo.base_dir, 'Backups')
+    def __backup_file(self, directory, release, commit=None):
+        backup_dir = os.path.join(self.__repo.base_dir, directory)
         if not os.path.isdir(backup_dir):
             os.mkdir(backup_dir)
-        file_name = f'{self.__repo.name}-{release}.sql'
-        return os.path.join(backup_dir, file_name)
+        file_name = f'{self.__repo.name}-{release}'
+        if commit:
+            file_name = f'{file_name}-{commit}'
+        return os.path.join(backup_dir, f'{file_name}.sql')
 
-    def __save_db(self):
+    def __save_db(self, release):
         """Save the database
         """
-        svg_file = self.__backup_file(self.previous)
+        commit = None
+        if self.__repo.production:
+            commit = self.__repo.hgit.last_commit()
+        svg_file = self.__backup_file('Backups', release, commit)
         print(f'Saving the database into {svg_file}')
         if os.path.isfile(svg_file):
             utils.error(
-                f"Oops! there is already a dump for the {self.previous} release.\n")
-            utils.error("Please remove it if you really want to proceed.\n", exit_code=1)
-        self.__repo.database.execute_pg_command('pg_dump', '-f', svg_file, stderr=subprocess.PIPE)
+                f"Oops! there is already a dump for the {release} release.\n")
+            utils.warning("Please remove it if you really want to proceed.\n")
+            sys.exit(1)
 
-    def __restore_previous_db(self):
+        self.__repo.database.execute_pg_command(
+            'pg_dump', '-f', svg_file, stderr=subprocess.PIPE)
+
+    def __restore_db(self, release):
         """Restore the database to the release_s version.
         """
-        print(f'Restoring the database to {self.previous}')
+        backup_file = self.__backup_file('Backups', release)
+        if not os.path.exists(backup_file):
+            raise FileNotFoundError(backup_file)
+        print(f'{utils.Color.green("Restoring the database to")} {utils.Color.bold(release)}')
         self.__repo.model.disconnect()
         self.__repo.database.execute_pg_command('dropdb')
         self.__repo.database.execute_pg_command('createdb')
         self.__repo.database.execute_pg_command(
-            'psql', '-f', self.__backup_file(self.previous), stdout=subprocess.DEVNULL)
+            'psql', '-f', backup_file, stdout=subprocess.DEVNULL)
         self.__repo.model.ping()
+
+    def __restore_previous_release(self):
+        db_release = self.__repo.database.last_release_s
+        self.__restore_db(db_release)
+        os.remove(self.__backup_file('Backups', db_release))
+        sys.exit(1)
 
     def __execute_sql(self, file_):
         "Execute sql query contained in sql file_"
@@ -165,47 +191,40 @@ class Patch:
             return
         try:
             self.__repo.model.execute_query(query)
-        except (psycopg2.Error, psycopg2.OperationalError, psycopg2.InterfaceError) as err:
+        except psycopg2.Error as err:
             utils.error(f'Problem with query in {file_.name}\n{err}\n')
-            self.__restore_previous_db()
-            os.remove(self.__backup_file(self.previous))
-            sys.exit(1)
+            self.__restore_previous_release()
 
     def __execute_script(self, file_):
         try:
+            python_path = os.environ.get('PYTHONPATH')
+            if python_path:
+                python_path = python_path.split(':')
+            else:
+                python_path = []
+            if self.__repo.base_dir:
+                os.environ.update({'PYTHONPATH': ':'.join([self.__repo.base_dir] + python_path)})
             subprocess.run(
                 ['python', file_.path],
-                env=os.environ.update({'PYTHONPATH': self.__repo.base_dir}),
+                cwd=self.__repo.base_dir,
+                env=os.environ,
                 shell=False, check=True)
         except subprocess.CalledProcessError as err:
             utils.error(f'Problem with script {file_}\n{err}\n')
-            self.__restore_previous_db()
-            os.remove(self.__backup_file(self.previous))
-            sys.exit(1)
+            self.__restore_previous_release()
 
-    def apply(self, release, force=False):
-        "Apply the patch in 'path'"
-        changelog_msg = ''
-        if self.__check_apply_or_re_apply() == 'apply':
-            self.__save_db()
-        else:
-            if not force:
-                okay = input('Do you want to re-apply the patch [y/N]?') or 'y'
-                if okay.upper() != 'Y':
-                    sys.exit()
-            self.__restore_previous_db()
-        print(f'Applying patch {release}')
-        files = []
-        major, minor, patch = release.split('.')
-        path = os.path.join(self.__patches_base_dir, major, minor, patch)
-        for file_ in os.scandir(path):
-            files.append({'name': file_.name, 'file': file_})
-        for elt in pydash.order_by(files, ['name']):
-            file_ = elt['file']
+    def __apply(self, path):
+        if not os.path.exists(path):
+            utils.warning(f"{path} does not exist. Skipping.\n")
+            sys.stderr.flush()
+            return
+        files_d = {elt.name: elt for elt in os.scandir(path)}
+        file_names = list(files_d.keys())
+        file_names.sort()
+        for file_name in file_names:
+            file_ = files_d[file_name]
             extension = file_.name.split('.').pop()
-            if file_.name == 'MANIFEST.py':
-                changelog_msg = json.loads(utils.read(file_))['changelog_msg']
-            if (not file_.is_file() or not (extension in ['sql', 'py'])):
+            if not (file_.is_file() and extension in ['sql', 'py']):
                 continue
             print(f'+ {file_.name}')
 
@@ -213,42 +232,115 @@ class Patch:
                 self.__execute_sql(file_)
             elif extension == 'py':
                 self.__execute_script(file_)
-        modules.generate(self.__repo)
+
+    def apply(self, release, force=False, save_db=True):
+        """Apply the release in 'path'
+
+        The history is first rebased on hop_main
+        """
+        if self.__repo.hgit.repos_is_clean():
+            self.__repo.hgit.rebase('hop_main')
+        db_release = self.__repo.database.last_release_s
+        changelog_msg = ''
+        if self.__check_apply_or_re_apply() == 'apply' and save_db:
+            self.__save_db(db_release)
+        elif not self.__repo.production:
+            if not force:
+                okay = input(f'Do you want to re-apply the release {release} [y/N]?') or 'y'
+                if okay.upper() != 'Y':
+                    sys.exit()
+            self.__restore_db(self.previous(db_release, 1))
+        app_upg = utils.Color.green('Upgrading to') if self.__repo.production else utils.Color.bold('Applying')
+        major, minor, patch = release.split('.')
+        print(utils.Color.green("Pre patch:"))
+        self.__apply(os.path.join(self.__patches_base_dir, 'pre'))
+        print(f'{app_upg} {utils.Color.green(release)}:')
+        self.__apply(os.path.join(self.__patches_base_dir, major, minor, patch))
+        if not self.__repo.production:
+            modules.generate(self.__repo)
+        print(utils.Color.green("Post patch:"))
+        self.__apply(os.path.join(self.__patches_base_dir, 'post'))
         self.__repo.database.register_release(major, minor, patch, changelog_msg)
 
     @property
     def state(self):
         "The state of a patch"
-        return '[Patch]'
+        if not self.__repo.devel:
+            return 'This repo is not in developement mode.'
+        if not self.__repo.production:
+            resp = ['[Releases in development]']
+            if len(self.__changelog.releases_in_dev) == 0:
+                resp.append("No release in development.\nUse `hop prepare`.")
+            for release in self.__changelog.releases_in_dev:
+                resp.append(f'- {release} (branch hop_{release})')
+        else:
+            resp = ['[Releases to apply]']
+            if len(self.__changelog.releases_to_apply_in_prod) == 0:
+                resp.append("No new release to apply.")
+            for release in self.__changelog.releases_to_apply_in_prod:
+                resp.append(f'- {release}')
+        return '\n'.join(resp)
+
 
     def undo(self, database_only=False):
         "Undo a patch."
-        self.__restore_previous_db()
+        db_release = self.__repo.database.last_release_s
+        previous_release = self.previous(db_release, 1)
+        self.__restore_db(previous_release)
         if not database_only:
             modules.generate(self.__repo)
-        os.remove(self.__backup_file(self.previous))
+        os.remove(self.__backup_file('Backups', previous_release))
+
+    def sync_package(self):
+        "Synchronise the package with the current database model"
+        modules.generate(self.__repo)
 
     def release(self, push):
         "Release a patch"
+        # We must be on the first branch in devel (see CHANGELOG)
+        next_release = self.__repo.changelog.releases_in_dev[0]
+        next_branch = f'hop_{next_release}'
+        if next_branch != self.__repo.hgit.branch:
+            utils.error(f'Next release is {next_release} Please switch to the branch {next_branch}!\n', 1)
         # Git repo must be clean
         if not self.__repo.hgit.repos_is_clean():
             utils.error(
-                'Please `git commit` your changes before releasing the patch.\n', exit_code=1)
+                f'Please `git commit` your changes before releasing {next_release}.\n', exit_code=1)
         # The patch must be applied and the last to apply
-        if not self.__repo.database.last_release_s == self.last:
-            utils.error('Please `hop apply-patch` before releasing the patch.\n', exit_code=1)
+        if self.__repo.database.last_release_s != next_release:
+            utils.error(f'Please `hop test` before releasing {next_release}.\n', exit_code=1)
         # If we undo the patch (db only) and re-apply it the repo must still be clear.
         self.undo(database_only=True)
-        self.apply(self.last, force=True)
+        self.apply(next_release, force=True)
         if not self.__repo.hgit.repos_is_clean():
             utils.error(
-                'Something has changed when re-applying the patch. This should not happen.\n',
+                'Something has changed when re-applying the release. This should not happen.\n',
                 exit_code=1)
-        # the tests must pass
-        try:
-            subprocess.run(['pytest', self.__repo.name], check=True)
-        except subprocess.CalledProcessError:
-            utils.error('Tests must pass in order to release the patch.\n', exit_code=1)
-        # So far, so good
-        self.__repo.hgit.rebase_to_hop_main(push)
-        self.__update_changelog(self.__repo.database.last_release_s)
+        # do we have pytest
+        if PYTEST_OK:
+            try:
+                subprocess.run(['pytest', self.__repo.name], check=True)
+            except subprocess.CalledProcessError:
+                utils.error('Tests must pass in order to release.\n', exit_code=1)
+            # So far, so good
+            svg_file = self.__backup_file('Releases', next_release)
+            print(f'Saving the database into {svg_file}')
+            self.__repo.database.execute_pg_command(
+                'pg_dump', '-xO', '-f', svg_file, stderr=subprocess.PIPE)
+            self.__repo.hgit.add(svg_file)
+            self.__repo.hgit.commit("-m", f"Add sql for release {next_release}")
+            self.__repo.hgit.rebase_to_hop_main(push)
+        else:
+            utils.error('pytest is not installed!\n', 1)
+
+    def upgrade_prod(self):
+        "Upgrade the production"
+        self.__assert_main_branch()
+        self.__save_db(self.__repo.database.last_release_s)
+        for release in self.__repo.changelog.releases_to_apply_in_prod:
+            self.apply(release, save_db=False)
+
+    def restore(self, release):
+        "Restore the database and package to a release (in production)"
+        self.__restore_db(release)
+        # Do we have the backup
