+++ test/dummy_test.sh
@@ -2,70 +2,123 @@
 
 set -ex
 
-perl -spi -e 's=True=False=' ~/.halform/hop_test
+if [ -n "$GITHUB_ENV" ]
+then
+   git config --global user.email "half_orm_ci@collorg.org"
+   git config --global user.name "HalfORM CI"
+fi
 
-rm -rf hop_test
+cd -- "$( dirname -- "${BASH_SOURCE[0]}" )"
+HALFORM_DIR=$PWD/../../..
+CI_PROJECT_DIR=$HALFORM_DIR
+HALFORM_CONF_DIR=$PWD/.config
+
+cat $HALFORM_CONF_DIR/hop_test
+# switch to development mode
+perl -spi -e 's=True=False=' $HALFORM_CONF_DIR/hop_test
+
+clean_db() {
+    rm -rf hop_test
+    set +e
+    dropdb hop_test
+    set -e
+}
+
+pwd
+clean_db
+createdb hop_test
+yes | hop new hop_test --devel
+clean_db
+# it should be able to create a repo not in devel mode
+yes | hop new hop_test
+cd hop_test
+
+hop sync-package
 set +e
-dropdb hop_test
+# it should FAIL: prepare is only available in devel mode
+hop prepare
+if [ $? = 0 ]; then exit 1; fi
 set -e
 
-yes | hop new hop_test
-
+cd ..
+pwd
+clean_db
+# it should be able to create a repo in devel mode
+yes | hop new hop_test --devel
 cd hop_test
 
 tree .
 
-hop
-perl -spi -e 's=False=True=' ~/.halform/hop_test
-echo 'IN PRODUCTION'
-hop
-perl -spi -e 's=True=False=' ~/.halform/hop_test
+rm -rf /tmp/hop_test.git
 
+# We need a central repo to release and simulate a production environment
+git init --bare /tmp/hop_test.git
 
-git remote add origin git@gite.lirmm.fr:maizi/tmptest.git
-git push -uf origin hop_main
-
-hop prepare-patch -m "First patch release" << EOF
+# 0.0.1
+hop prepare -m "First patch release" << EOF
 patch
 EOF
+
+if [ `git branch --show-current` != 'hop_0.0.1' ] ; then echo "It should be on branch hop_0.0.1" ; exit 1 ; fi
+
 echo 'create table first ( a text primary key )' > Patches/0/0/1/a.sql
-hop apply-patch
+hop apply
 git add .
 git commit -m "First table"
-hop release-patch
+set +e
+which pytest
+if [ $? = 0 ]; then pip uninstall -y pytest ; fi
 
-hop prepare-patch -l patch -m "Second patch release"
+# It should FAIL: pytest must be install to commit release
+hop release
+if [ $? = 0 ]; then exit 1; fi
+set -e
+pip install pytest
+hop release
 
-echo 'create table a ( a text primary key )' > Patches/0/0/2/a.sql
-echo 'print("I am a script without x permission...")' > Patches/0/0/2/a.py
+# 0.0.2
+hop prepare -l patch -m "Second patch release"
+if [ `git branch --show-current` != 'hop_0.0.2' ] ; then echo "It should be on branch hop_0.0.2" ; exit 1 ; fi
 
-hop apply-patch
+echo 'create table a ( a text primary key, class int, "class + 1" int )' > Patches/0/0/2/00_a.sql
+cat > Patches/0/0/2/a.py << EOF
+from hop_test.public.first import First
+list(First())
+EOF
+
+# It should be able to apply multiple times
+sync; sync
+hop apply
 
 tree .
 
 hop
 
-yes | hop apply-patch
+yes | hop apply
 
-hop apply-patch -f
+hop apply
 
 git add .
 git commit -m "(wip) First"
 git status
 
-echo 'create table a ( a text primary key, bla text )' > Patches/0/0/2/a.sql
+# Change of the model + the patches must be executed in order
+echo 'create table a ( a text primary key, bla text )' > Patches/0/0/2/00_a.sql
+echo 'alter table a add column b text' > Patches/0/0/2/01_a.sql
+echo 'alter table a alter column b set not null' > Patches/0/0/2/02_a.sql
+echo 'alter table a add constraint b_uniq unique(b)' > Patches/0/0/2/03_a.sql
 
-hop undo-patch
+# hop undo
 
-hop apply-patch
-git diff hop_test/public/a.py
+hop apply
+# git diff hop_test/public/a.py
 
 git status
-git hist
 
 set +e
 # should commit before release
-hop release-patch -m "First release"
+hop release -m "First release"
+if [ $? = 0 ]; then exit 1; fi
 set -e
 
 git add .
@@ -74,7 +127,8 @@ git commit -m "(wip) ajout de a.bla"
 touch dirty
 set +e
 # git repo must be clean
-hop release-patch
+hop release
+if [ $? = 0 ]; then exit 1; fi
 set -e
 rm dirty
 
@@ -82,39 +136,35 @@ echo '        verybad' >> hop_test/public/a.py
 git add .
 git commit -m "(bad)"
 set +e
-# git repo must be clean
-hop release-patch
+# test must pass
+hop release # 0.0.2
+if [ $? = 0 ]; then exit 1; fi
 set -e
+git branch
 git reset HEAD~ --hard
 
-hop release-patch --push
+set +e
+# git repo must have an origin to push
+hop release --push
+if [ $? = 0 ]; then exit 1; fi
+set -e
+
+git remote add origin /tmp/hop_test.git
+git status
+
+sync; sync
+hop release --push # 0.0.2
 
 git status
 
-# echo 'APPLY PATCH IN PRODUCTION'
-# git checkout hop_0.0.1
-# hop undo-patch -d
-# git checkout hop_main
-
-# perl -spi -e 's=False=True=' ~/.halform/hop_test
-
-# hop
-
-# hop apply-patch
-
-# perl -spi -e 's=True=False=' ~/.halform/hop_test
-
-# mv ../../half_orm_packager/version.txt ../../half_orm_packager/version.txt.o
-
-# echo '0.1.0' > ../../half_orm_packager/version.txt
-
 hop
 
-hop prepare-patch -l minor -m "First minor patch"
+hop prepare -l minor -m "First minor patch" # 0.1.0
+if [ `git branch --show-current` != 'hop_0.1.0' ] ; then echo "It should be on branch hop_0.1.0" ; exit 1 ; fi
 
 echo 'create table b ( b text primary key, a text references a )' > Patches/0/1/0/b.sql
 
-hop apply-patch
+hop apply
 
 tree
 
@@ -124,60 +174,18 @@ git status
 git add .
 git commit -m "(wip) Second"
 
+echo add a_fk in B.Fkeys
 cat > hop_test/public/b.py << EOF
-# hop release: 0.1.0a0
-# pylint: disable=wrong-import-order, invalid-name, attribute-defined-outside-init
+from half_orm.field import Field
+from half_orm.fkey import FKey
+from hop_test import MODEL, ho_dataclasses
 
-"""The module hop_test.public.b povides the B class.
-
-WARNING!
-
-This file is part of the hop_test package. It has been generated by the
-command hop. To keep it in sync with your database structure, just rerun
-hop update.
-
-More information on the half_orm library on https://github.com/collorg/halfORM.
-
-DO NOT REMOVE OR MODIFY THE LINES BEGINING WITH:
-#>>> PLACE YOUR CODE BELOW...
-#<<< PLACE YOUR CODE ABOVE...
-
-MAKE SURE YOUR CODE GOES BETWEEN THESE LINES OR AT THE END OF THE FILE.
-hop ONLY PRESERVES THE CODE BETWEEN THESE MARKS WHEN IT IS RUN.
-"""
-
-from hop_test.db_connector import base_relation_class
-
-#>>> PLACE YOUR CODE BELLOW THIS LINE. DO NOT REMOVE THIS LINE!
+#>>> PLACE YOUR CODE BELOW THIS LINE. DO NOT REMOVE THIS LINE!
 import datetime
 #<<< PLACE YOUR CODE ABOVE THIS LINE. DO NOT REMOVE THIS LINE!
 
-__RCLS = base_relation_class('public.b')
-
-class B(__RCLS):
-    """
-    __RCLS: <class 'half_orm.model.Table_Hop_testPublicB'>
-    This class allows you to manipulate the data in the PG relation:
-    TABLE: "hop_test":"public"."b"
-    FIELDS:
-    - b: (text) NOT NULL
-    - a: (text)
-
-    PRIMARY KEY (b)
-    FOREIGN KEY:
-    - b_a_fkey: ("a")
-     ↳ "hop_test":"public"."a"(a)
-
-    To use the foreign keys as direct attributes of the class, copy/paste the Fkeys bellow in
-    your code as a class attribute and replace the empty string(s) key(s) with the alias you
-    want to use. The aliases must be unique and different from any of the column names. Empty
-    string keys are ignored.
-
-    Fkeys = {
-        '': 'b_a_fkey',
-    }
-    """
-    #>>> PLACE YOUR CODE BELLOW THIS LINE. DO NOT REMOVE THIS LINE!
+class B(MODEL.get_relation_class('public.b'), ho_dataclasses.DC_PublicB):
+    #>>> PLACE YOUR CODE BELOW THIS LINE. DO NOT REMOVE THIS LINE!
     Fkeys = {
         'a_fk': 'b_a_fkey',
     }
@@ -185,8 +193,10 @@ class B(__RCLS):
     def __init__(self, **kwargs):
         super().__init__(**kwargs)
 
-        #>>> PLACE YOUR CODE BELLOW THIS LINE. DO NOT REMOVE THIS LINE!
+        #>>> PLACE YOUR CODE BELOW THIS LINE. DO NOT REMOVE THIS LINE!
 EOF
+
+hop apply
 
 git add .
 git commit -m "(b) with Fkeys"
@@ -199,10 +209,117 @@ echo 'print ("coucou")' > hop_test/Api/coucou/__init__.py
 git add .
 git commit -m "Add API"
 
-hop release-patch
+hop release
+if [ `git branch --show-current` != 'hop_main' ] ; then echo "It should be on branch hop_main" ; exit 1 ; fi
 
 git status
 git push
 git tag
 
-git log
+# It should be able to prepare a major, a minor and a patch simultaneously.
+hop prepare -l major -m major # 1.0.0
+if [ `git branch --show-current` != 'hop_1.0.0' ] ; then echo "It should be on branch hop_1.0.0" ; exit 1 ; fi
+
+hop prepare -l minor -m minor # 0.2.0
+if [ `git branch --show-current` != 'hop_0.2.0' ] ; then echo "It should be on branch hop_0.2.0" ; exit 1 ; fi
+
+touch Patches/0/2/0/coucou
+git add .
+git commit -m "[WIP] 0.2.0 test"
+hop apply
+
+echo toto > toto
+git status
+set +e
+# It should FAIL: the git repo is not clean
+hop prepare -l patch -m patch
+if [ $? = 0 ]; then echo "It should FAIL: the git repo is not clean"; exit 1; fi
+set -e
+
+
+git add toto
+git commit -m "toto"
+
+# Check rebase fails
+
+git checkout hop_main
+echo tata > toto
+git add .
+git commit -m "toto on hop_main"
+
+set +e
+# It should FAIL: hop_0.2.0 can't be rebased on hop_main
+hop
+if [ $? = 0 ]; then exit 1; fi
+set -e
+git branch
+git reset HEAD~ --hard
+
+git checkout hop_0.2.0
+
+# hop undo
+# git checkout hop_main
+#  (hop prepare should undo the changes and switch to hop_main branch)
+hop prepare -l patch -m patch # 0.1.1
+git branch
+
+set +e
+# It should FAIL: there is already on minor version in preparation.
+hop prepare -l minor -m coucou
+if [ $? = 0 ]; then exit 1; fi
+set -e
+
+hop
+
+git checkout hop_0.2.0
+set +e
+# It should FAIL: the patch in preparation must be released first.
+hop release
+if [ $? = 0 ]; then exit 1; fi
+set -e
+if [ ! -f Backups/hop_test-0.1.0.sql ]; then exit 1; fi
+rm Backups/hop_test-0.1.0.sql
+git checkout hop_0.1.1
+hop apply
+touch Patches/0/1/1/coucou
+git add .
+git commit -m "[WIP] 0.1.1 test"
+hop apply
+hop release
+
+cat > TODO << EOF
+something
+EOF
+git add TODO
+git commit -m "Add todo to check rebase on apply"
+
+git checkout hop_0.2.0
+# It should rebase hop_0.2.0 on hop_main (hop_0.1.1 has been released)
+hop apply
+hop release
+
+hop prepare -l patch -m "0.2.1..." # 0.2.1
+hop apply
+touch Patches/0/2/1/coucou
+git add .
+git commit -m "[0.2.1] coucou"
+hop release
+
+git checkout hop_1.0.0
+hop apply
+touch Patches/1/0/0/coucou
+git add .
+git commit -m "[WIP] 1.0.0 test"
+hop apply
+hop release
+
+git push
+
+echo 'APPLY PATCH IN PRODUCTION'
+perl -spi -e 's=False=True=' $HALFORM_CONF_DIR/hop_test
+hop
+hop restore 0.0.1
+hop
+# It should upgrade to the latest version released (1.0.0)
+hop upgrade
+perl -spi -e 's=True=False=' $HALFORM_CONF_DIR/hop_test
